+++ src/libs/cabac/context_selector.h
@@ -17,17 +17,17 @@ namespace contextSelector{
 
     //void getContextIdsOrder1BI(std::vector<unsigned int>&, unsigned int, unsigned int, unsigned int);
 
-    unsigned int getContextIdOrder1TU(unsigned int, unsigned int, unsigned int);
+    unsigned int getContextIdOrder1TU(unsigned int, uint64_t, unsigned int);
 
-    void getContextIdsOrder1TU(std::vector<unsigned int>&, unsigned int, unsigned int, unsigned int);
+    void getContextIdsOrder1TU(std::vector<unsigned int>&, uint64_t, unsigned int, unsigned int);
 
-    unsigned int getContextIdOrder1EG0(unsigned int, unsigned int, unsigned int);
+    unsigned int getContextIdOrder1EG0(unsigned int, uint64_t, unsigned int);
 
-    void getContextIdsOrder1EG0(std::vector<unsigned int>&, unsigned int, unsigned int, unsigned int);
+    void getContextIdsOrder1EG0(std::vector<unsigned int>&, uint64_t, unsigned int, unsigned int);
 
-    unsigned int getContextIdOrder1EGk(unsigned int, unsigned int, unsigned int, unsigned int);
+    unsigned int getContextIdOrder1EGk(unsigned int, uint64_t, unsigned int, unsigned int);
 
-    void getContextIdsOrder1EGk(std::vector<unsigned int>&, unsigned int, unsigned int, unsigned int, unsigned int);
+    void getContextIdsOrder1EGk(std::vector<unsigned int>&, uint64_t, unsigned int, unsigned int, unsigned int);
 };
 
 #endif  // RWTH_PYTHON_IF
